PAUSE now pauses the game in any mode.  (Which may be overkill for menu modes.)
It should now be possible to make proper screenshots.  It is now also easier
to and see the momentary timings you get with F12 (but ignore the frame rate
and the penultimate 'other' entry -- there's a fixed delay of 40ms per frame
in skip_a_tick.  Actually that should be a 50ms delay... Strange.